+++ retrocache/src/main/java/com/andiag/retrocache/annotations/Caching.java
@@ -6,13 +6,13 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * Created by Canalejas on 02/02/2017.
+ * Created by IagoCanalejas on 02/02/2017.
  * Annotation used to avoid a {@link retrofit2.http.GET} method to be cached.
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.METHOD)
 public @interface Caching {
 
-    boolean active() default true;
+    boolean enabled() default true;
 
 }