+++ retrocache/src/main/java/com/andiag/retrocache/utils/ByteArraySerializer.java
@@ -5,7 +5,7 @@ import com.iagocanalejas.dualcache.interfaces.Serializer;
 import java.nio.charset.Charset;
 
 /**
- * Created by Canalejas on 12/01/2017.
+ * Created by IagoCanalejas on 12/01/2017.
  * Basic parser that encode and decode byte[] matching default charset
  */
 public class ByteArraySerializer implements Serializer<byte[]> {